pyproject.toml
[project]
name = "nano-hairstyle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "fastapi>=0.117.1",
    "google-auth==2.40.3",
    "google-genai==1.38.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "pillow==11.3.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests==2.32.5",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "tenacity==9.1.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn>=0.36.0",
    "websockets==15.0.1",
]
